improve alidropshopping and develop like preview and picker

+++ database/migrations/2021_02_24_154706_add_deal_type_to_flash_deals.php
@@ -13,9 +13,11 @@ class AddDealTypeToFlashDeals extends Migration
      */
     public function up()
     {
-        Schema::table('flash_deals', function (Blueprint $table) {
-            $table->string('deal_type')->nullable();
-        });
+        if (! Schema::hasColumn('flash_deals', 'deal_type')) {
+            Schema::table('flash_deals', function (Blueprint $table) {
+                $table->string('deal_type')->nullable();
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class AddDealTypeToFlashDeals extends Migration
      */
     public function down()
     {
-        Schema::table('flash_deals', function (Blueprint $table) {
-            $table->dropColumn(['deal_type']);
-        });
+        if (Schema::hasColumn('flash_deals', 'deal_type')) {
+            Schema::table('flash_deals', function (Blueprint $table) {
+                $table->dropColumn(['deal_type']);
+            });
+        }
     }
 }
